Automatic train an Echo State Network

This function trains an Echo State Network (ESN) to a univariate time series.

auto_esn(
  .data,
  specials,
  const = TRUE,
  lags = NULL,
  n_fourier = NULL,
  n_initial = 10,
  n_res = 200,
  n_seed = 42,
  alpha = 0.8,
  rho = 1,
  lambda = 1,
  density = 0.1,
  scale_inputs = c(-1, 1),
  scale_runif = c(-0.5, 0.5),
  inf_crit = "bic",
  control_inputs = list(tune = TRUE, n_sample = 1000),
  control_pars = list(tune = TRUE, lower = c(0, 0.1, 0.1), upper = c(1, 5, 10)),
  ...
)

Arguments

.data
A tsibble containing the time series data.

specials
Currently not is use.

const
Logical value. If TRUE, a constant term (intercept) is used.

lags
A list containing integer vectors with the lags associated with each input variable.

n_fourier
Integer vector. The number of fourier terms (seasonal cycles per period).

n_initial
Integer value. The number of observations of internal states for initial drop out (throw-off).

n_res
Integer value. The number of internal states within the reservoir (hidden layer).

n_seed
Integer value. The seed for the random number generator (for reproducibility).

alpha
Numeric value. The leakage rate (smoothing parameter) applied to the reservoir.

rho
Numeric value. The spectral radius for scaling the reservoir weight matrix.

lambda
Numeric value. The regularization (shrinkage) parameter for ridge regression.

density
Numeric value. The connectivity of the reservoir weight matrix (dense or sparse).

scale_inputs
Numeric vector. The lower and upper bound for scaling the time series data.

scale_runif
Numeric vector. The lower and upper bound of the uniform distribution.

inf_crit
Character value. The information criterion inf_crit = c("aic", "bic", "hq").

control_inputs
A list containing control values for the automatic selection of model inputs:

tune: Logical value. If TRUE, the model inputs are tuned, otherwise model inputs are used as defined.

n_sample: Integer value. The number of samples for the random grid.

control_pars
A list containing control values for the hyperparameter tuning:

tune: Logical value. If TRUE, the hyperparameters are tuned, otherwise hyperparameters are used as defined.

lower: Numeric vector. The lower bounds for alpha, rho and lambda used in the optimization.

upper: Numeric vector. The upper bounds for alpha, rho and lambda used in the optimization.

...
Further arguments passed to stats::optim()

Value

An object of class ESN.