Estimate information criterion for hyperparameter tuning

The function tune_pars() estimates the information criterion for hyperparameter tuning. The function returns a numeric value (= information criterion), which is then minimized within a call to optim() for varying hyperparameters.

tune_pars(
  data,
  pars,
  const,
  lags,
  n_fourier,
  period,
  n_diff,
  n_res,
  n_initial,
  n_seed,
  density,
  inf_crit,
  scale_inputs,
  scale_runif
)

Arguments

data
A tsibble containing the time series data.

pars
Numeric vector containing the hyperparameters.

const
Logical value. If TRUE, a constant term (intercept) is used.

lags
A list containing integer vectors with the lags associated with each input variable.

n_fourier
Integer vector. The number of fourier terms (seasonal cycles per period).

period
Integer vector. The periodicity of the time series (e.g. period = c(12) for monthly data or period = c(24, 168) for hourly data).

n_diff
Integer value. The number of differences.

n_res
Integer value. The number of internal states within the reservoir (hidden layer).

n_initial
Integer value. The number of observations of internal states for initial drop out (throw-off).

n_seed
Integer value. The seed for the random number generator (for reproducibility).

density
Numeric value. The connectivity of the reservoir weight matrix (dense or sparse).

inf_crit
Character value. The information criterion inf_crit = c("aic", "bic", "hq").

scale_inputs
Numeric vector. The lower and upper bound for scaling the time series data.

scale_runif
Numeric vector. The lower and upper bound of the uniform distribution.

Value

model_value Numeric value. The information criterion to be minimized.