Train an Echo State Network

This function trains an Echo State Network (ESN) to a univariate time series.

train_esn(
  data,
  lags,
  n_fourier,
  period,
  const = TRUE,
  n_diff = 0,
  n_res = 200,
  n_initial = 10,
  n_seed = 42,
  rho = 0.7,
  alpha = 0.85,
  lambda = 12,
  density = 0.1,
  weights = NULL,
  scale_runif = c(-0.5, 0.5),
  scale_inputs = c(-1, 1)
)

Arguments

data
A tsibble containing the time series data.

lags
A list containing integer vectors with the lags associated with each input variable.

n_fourier
Integer vector. The number of fourier terms (seasonal cycles per period).

period
Integer vector. The periodicity of the time series (e.g. for monthly data period = c(12), for hourly data period = c(24, 168)).

const
Logical value. If TRUE, a constant term (intercept) is used.

n_diff
Integer vector. The number of non-seasonal differences.

n_res
Integer value. The number of internal states within the reservoir (hidden layer).

n_initial
Integer value. The number of observations of internal states for initial drop out (throw-off).

n_seed
Integer value. The seed for the random number generator (for reproducibility).

rho
Numeric value. The spectral radius for scaling the reservoir weight matrix.

alpha
Numeric value. The leakage rate (smoothing parameter) applied to the reservoir.

lambda
Numeric value. The regularization (shrinkage) parameter for ridge regression.

density
Numeric value. The connectivity of the reservoir weight matrix (dense or sparse).

scale_runif
Numeric vector. The lower and upper bound of the uniform distribution.

scale_inputs
Numeric vector. The lower and upper bound for scaling the time series data.

Value

A list containing:

data: The original input data as tsibble.

actual: A tsibble containing the actual values.

fitted: A tsibble containing the fitted values.

error: A tsibble containing the residuals (i.e. actual - fitted).

states_train: A tsibble containing the internal states.

method: A list containing several objects and information of the trained ESN (weight matrices, hyperparameters, model metrics, etc.).